+++ ethereum/spec/src/main/java/tech/pegasys/teku/spec/datastructures/networking/libp2p/rpc/DataColumnSidecarsByRootRequestMessage.java
@@ -38,11 +38,11 @@ public class DataColumnSidecarsByRootRequestMessage extends SszListImpl<DataColu
   @Override
   public int getMaximumResponseChunks() {
     return stream()
-            .reduce(
-                    0,
-                    (current, dataColumnsByRootIdentifier) ->
-                            current + dataColumnsByRootIdentifier.getColumns().size(),
-                    Integer::sum);
+        .reduce(
+            0,
+            (current, dataColumnsByRootIdentifier) ->
+                current + dataColumnsByRootIdentifier.getColumns().size(),
+            Integer::sum);
   }
 
   @Override