+++ teku/src/test/java/tech/pegasys/teku/cli/options/ValidatorOptionsTest.java
@@ -123,22 +123,6 @@ public class ValidatorOptionsTest extends AbstractBeaconNodeCommandTest {
   }
 
   @Test
-  public void shouldEnableBlindedBeaconBlocks() {
-    final String[] args = {"--validators-proposer-blinded-blocks-enabled", "true"};
-    final TekuConfiguration config = getTekuConfigurationFromArguments(args);
-    assertThat(config.validatorClient().getValidatorConfig().isBlindedBeaconBlocksEnabled())
-        .isTrue();
-  }
-
-  @Test
-  public void shouldNotUseBlindedBeaconBlocksByDefault() {
-    final String[] args = {};
-    final TekuConfiguration config = getTekuConfigurationFromArguments(args);
-    assertThat(config.validatorClient().getValidatorConfig().isBlindedBeaconBlocksEnabled())
-        .isFalse();
-  }
-
-  @Test
   public void shouldSetValidatorRegistrationTimestampOverride() {
     final String[] args = {"--Xvalidators-builder-registration-timestamp-override", "120000"};
     final TekuConfiguration config = getTekuConfigurationFromArguments(args);